Refactor Intake classes
